Manually Merged Player and Swordsman scenes/scripts from Jeromes Branch
Finished working implementation of the pushable boxes and manually combined the swordsman scenes/ scripts of Jerome's branch.

Box notes: Because of the intended behavior of the pushable boxes, rigidbody2d did not give the intended result. Switching to characterbody2d and instead relying on raycasting to detect collision/ push the box  rather than a push area with rigidbody2d, fixed the behavior. As the boxes are to be used for puzzles, I needed them to be strict to the grid and not too free form/moving. This helps future box puzzle/level designs.